Test: Refactor AuthStore into services
