+++ frontend/src/lib/components/common/SignIn.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { authStore } from "$lib/stores/auth.store";
+  import { signIn as signInService } from "$lib/services/auth.services";
   import { i18n } from "$lib/stores/i18n";
   import { toastsError } from "$lib/stores/toasts.store";
   import { Spinner } from "@dfinity/gix-components";
@@ -14,7 +14,7 @@
         err,
       });
 
-    await authStore.signIn(onError);
+    await signInService(onError);
   };
 
   let disabled = true;
