+++ frontend/src/lib/services/auth.services.ts
@@ -1,16 +1,44 @@
+import { resetAgents } from "$lib/api/agent.api";
+import {
+  AUTH_SESSION_DURATION,
+  IDENTITY_SERVICE_URL,
+  OLD_MAINNET_IDENTITY_SERVICE_URL,
+} from "$lib/constants/identity.constants";
+import { NNS_IC_APP_DERIVATION_ORIGIN } from "$lib/constants/origin.constants";
 import { authStore } from "$lib/stores/auth.store";
 import { startBusy } from "$lib/stores/busy.store";
 import { toastsShow } from "$lib/stores/toasts.store";
 import type { ToastMsg } from "$lib/types/toast";
+import { createAuthClient } from "$lib/utils/auth.utils";
+import { isNnsAlternativeOrigin } from "$lib/utils/env.utils";
 import { replaceHistory } from "$lib/utils/route.utils";
 import type { Identity } from "@dfinity/agent";
 import { AnonymousIdentity } from "@dfinity/agent";
+import type { AuthClient } from "@dfinity/auth-client";
 import type { ToastLevel } from "@dfinity/gix-components";
 import { get } from "svelte/store";
 
 const msgParam = "msg";
 const levelParam = "level";
 
+// We have to keep the authClient object in memory because calling the `authClient.login` feature should be triggered by a user interaction without any async callbacks call before calling `window.open` to open II
+// @see agent-js issue [#618](https://github.com/dfinity/agent-js/pull/618)
+let authClient: AuthClient | undefined | null;
+
+const getIdentityProvider = () => {
+  // If we are in mainnet in the old domain, we use the old identity provider.
+  if (location.host === "nns.ic0.app") {
+    return OLD_MAINNET_IDENTITY_SERVICE_URL;
+  }
+
+  return IDENTITY_SERVICE_URL;
+};
+
+/**
+ * Call auth-client log out and set null in the store. Started with a user interaction ("click on a button")
+ *
+ * note: clearing idb auth keys does not happen in the state management but afterwards in its caller function (see <Logout/>)
+ */
 export const logout = async ({
   msg = undefined,
 }: {
@@ -19,7 +47,18 @@ export const logout = async ({
   // To mask not operational UI (a side effect of sometimes slow JS loading after window.reload because of service worker and no cache).
   startBusy({ initiator: "logout" });
 
-  await authStore.signOut();
+  const client: AuthClient = authClient ?? (await createAuthClient());
+
+  await client.logout();
+
+  resetAgents();
+
+  // We currently do not have issue because the all screen is reloaded after sign-out.
+  // But, if we wouldn't, then agent-js auth client would not be able to process next sign-in if object would be still in memory with previous partial information. That's why we reset it.
+  // This fix a "sign in -> sign out -> sign in again" flow without window reload.
+  authClient = null;
+
+  authStore.setNoIdentity();
 
   if (msg) {
     appendMsgToUrl(msg);
@@ -32,6 +71,41 @@ export const logout = async ({
 
   // We reload the page to make sure all the states are cleared
   window.location.reload();
+};
+
+/**
+ * Log in flow. Started with a user interaction ("click on a button")
+ */
+export const signIn = async (onError: (error?: string) => void) => {
+  authClient = authClient ?? (await createAuthClient());
+
+  await authClient?.login({
+    identityProvider: getIdentityProvider(),
+    ...(isNnsAlternativeOrigin() && {
+      derivationOrigin: NNS_IC_APP_DERIVATION_ORIGIN,
+    }),
+    maxTimeToLive: AUTH_SESSION_DURATION,
+    onSuccess: () => {
+      authStore.setIdentity(authClient?.getIdentity());
+    },
+    onError,
+  });
+};
+
+/**
+ * Sync Auth: query auth-client to get the status of the authentication
+ * a. if authenticated only, set identity in the global state
+ * b. if not authenticated, set null in store
+ *
+ * the sync function is performed when the app boots and on any change in the local storage (see <Guard/>)
+ *
+ * note: auth-client is initialized with an anonymous principal. By querying "isAuthenticated", the library checks for a valid chain and also that the principal is not anonymous.
+ */
+export const syncAuth = async () => {
+  authClient = authClient ?? (await createAuthClient());
+  const isAuthenticated = await authClient.isAuthenticated();
+
+  authStore.setIdentity(isAuthenticated ? authClient.getIdentity() : null);
 };
 
 /**
